Even more changes...
- Removed 3DoF mode from multiplayer (it would never be used)
- Prevent player clipping though solid objects/walls in Multiplayer
- vr_weaponPitch - Allows adjustment of weapon pitch - default: -20
- vr_heightAdjust - Additional height adjustment in metres: default: 0.0
- vr_twoHandedWeapons - Allows disabling of two handed weapon mode, set to 0 to disable

+++ code/vr/vr_renderer.c
@@ -244,7 +244,7 @@ void VR_DrawFrame( engine_t* engine ) {
 	const ovrMatrix4f projectionMatrix = ovrMatrix4f_CreateProjectionFov(
 			fov_x, fov_y, 0.0f, 0.0f, 1.0f, 0.0f );
 
-	static int playerYaw = 0;
+	static float playerYaw = 0;
 
     int eyeW, eyeH;
     VR_GetResolution(engine, &eyeW, &eyeH);
